split up control_surfaces_test
