+++ src/bin/test/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(loratest)
-add_subdirectory(control_surfaces)
+add_subdirectory(control_surfaces_test_bot)
+add_subdirectory(control_surfaces_test_box)
