+++ src/bin/test/control_surfaces_test_bot/CMakeLists.txt
@@ -1,0 +1,18 @@
+set(APP jaiabot_control_surfaces_test_bot)
+
+protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS ${CMAKE_CURRENT_BINARY_DIR} config.proto)
+
+add_executable(${APP}
+  app.cpp
+  ${PROTO_SRCS} ${PROTO_HDRS})
+
+target_link_libraries(${APP}
+  goby
+  goby_zeromq
+  jaiabot_messages)
+
+if(export_goby_interfaces)
+  generate_interfaces(${APP})
+endif()
+
+project_install_bin(${APP})
