+++ src/bin/test/control_surfaces_test_bot/config.proto
